view/requirements.txt
fastapi
argon2-cffi
cryptography
python-dotenv
hashlib
secrets
mysql-connector-python